requirements.txt
Django==1.9.5
django-bootstrap-form==3.2.1
django-crispy-forms==1.6.1
Pillow==3.4.2

